requirements.txt
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==2.1.0
psycopg2==2.7.3.2
pytz==2019.1
sqlparse==0.3.0
django-rest-auth==0.9.5
